removed lzz
