+++ .gitattributes
@@ -1,5 +1,1 @@
 *.lzz linguist-language=C++
-*.cpp -diff
-*.hpp -diff
-*.c -diff
-*.h -diff
