+++ src/util/query-macros.cpp
